+++ src/main/java/com/GymManager/Backend/domain/service/AccessLogService.java
@@ -7,7 +7,7 @@ import com.GymManager.Backend.persistence.entity.AccessLogEntity;
 import java.util.List;
 
 public interface AccessLogService {
-    Boolean save(Integer userId);
+    Boolean save(Long userId);
     List<AccessLogEntity> getAllMemberToday();
     List<AccessLogEntity> getAllMemberMonth();
     void deleteAllToday();
